Separated local storage.
